# apps/dashboard/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flowstate API",
    "description": "Property valuation and underwriting API for real estate investors. Get property data, comparables, ARV calculations, and investment recommendations.\n\n## Authentication\n\nThe API supports two authentication methods:\n- **Bearer Token**: For API access via `/v1/*` endpoints. Use API keys from your dashboard.\n- **Session Cookie**: For dashboard endpoints (`/user/*`, `/appraisal-presets/*`). Managed via Better Auth.\n\n## Real-time Analysis\n\nProperty analysis is async. POST to `/v1/analyze` returns a job ID immediately. Get results via:\n- **WebSocket**: `/ws/analyze/:jobId` for real-time streaming\n- **Polling**: `GET /v1/analyze/jobs/:jobId` for status checks",
    "version": "1.0.0",
    "contact": {
      "name": "Flowstate Support",
      "url": "https://flowstate.homes"
    }
  },
  "servers": [
    {
      "url": "https://api.flowstate.homes",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787",
      "description": "Local Development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/analyze": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Start property analysis",
        "description": "Start an async property analysis using Cloudflare Workflows. Returns immediately with a job ID and URLs for status/streaming.\n\nThe analysis includes:\n1. Property lookup and data fetching\n2. Comparable sales search\n3. Photo fetching and AI vision analysis\n4. Appraisal adjustments\n5. ARV calculation and investment metrics\n\nResults can be retrieved via SSE streaming or HTTP polling.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalyzeRequest" },
              "examples": {
                "simple": {
                  "summary": "Simple address lookup",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607"
                  }
                },
                "with_photos": {
                  "summary": "Analysis with photo analysis",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607",
                    "photoAnalysis": {
                      "enabled": true,
                      "maxComps": 10
                    }
                  }
                },
                "advanced": {
                  "summary": "Advanced analysis with all options",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607",
                    "searchOptions": {
                      "radiusMiles": 1,
                      "maxComps": 15,
                      "monthsBack": 12
                    },
                    "appraisalRules": {
                      "filters": [
                        { "type": "sale_age", "enabled": true, "value": 12 },
                        { "type": "sqft_diff", "enabled": true, "value": 25 }
                      ],
                      "adjustments": [
                        { "type": "bedroom", "enabled": true, "amount": 5000 }
                      ]
                    },
                    "buybox": {
                      "rehabLevelIndex": 2,
                      "closingCostsPercent": 10,
                      "carryingCostsPercent": 5,
                      "wholesaleFee": 10000
                    },
                    "enrichment": {
                      "permits": true,
                      "floodZone": true
                    },
                    "photoAnalysis": {
                      "enabled": true,
                      "maxComps": 10,
                      "requireBetterOrEqual": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Analysis job queued successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QueueJobResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/jobs/{jobId}": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get job status",
        "description": "Get the current status of an analysis job. Returns job state, progress, and result (if completed).",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "description": "Job ID returned from POST /v1/analyze",
            "schema": { "type": "string" }
          },
          {
            "name": "propertyKey",
            "in": "query",
            "required": true,
            "description": "Property key returned from POST /v1/analyze",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobStatusResponse" }
              }
            }
          },
          "400": {
            "description": "Missing propertyKey parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/stream-token": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Generate stream token",
        "description": "Generate a short-lived signed token for SSE stream authentication. The token expires in 5 minutes and is scoped to a specific job.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["jobId", "propertyKey"],
                "properties": {
                  "jobId": { "type": "string", "description": "Job ID from POST /v1/analyze" },
                  "propertyKey": { "type": "string", "description": "Property key from POST /v1/analyze" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream token generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "description": "Signed JWT token" },
                        "streamUrl": { "type": "string", "description": "Full SSE stream URL with token" },
                        "expiresIn": { "type": "integer", "example": 300, "description": "Token lifetime in seconds" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/defaults": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get default parameters",
        "description": "Get default values for search options, appraisal rules, buybox parameters, and enrichment settings.",
        "responses": {
          "200": {
            "description": "Default parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnalyzeDefaultsResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from your dashboard (e.g., `fs_abc123...`)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie from Better Auth (used for dashboard endpoints)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Property not found" },
          "code": { "type": "string", "example": "NOT_FOUND" }
        }
      },
      "AnalyzeRequest": {
        "type": "object",
        "description": "Request body for property analysis. Provide either `address`, `streetAddress` + city/state/zip, or `propertyId`.",
        "properties": {
          "address": {
            "type": "string",
            "description": "Full property address",
            "example": "123 Main St, Tampa, FL 33607"
          },
          "streetAddress": { "type": "string", "description": "Street address only" },
          "city": { "type": "string", "description": "City name" },
          "state": { "type": "string", "description": "State code (e.g., FL)" },
          "zipCode": { "type": "string", "description": "5-digit ZIP code" },
          "propertyId": { "type": "string", "description": "Property ID from a previous lookup" },
          "searchOptions": {
            "type": "object",
            "properties": {
              "radiusMiles": { "type": "number", "default": 1, "description": "Search radius for comparables" },
              "maxComps": { "type": "integer", "default": 10, "description": "Maximum comparables to return" },
              "monthsBack": { "type": "integer", "default": 12, "description": "How far back to search for sales" }
            }
          },
          "appraisalRules": {
            "type": "object",
            "description": "Custom appraisal rules",
            "properties": {
              "filters": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AppraisalFilter" }
              },
              "adjustments": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AppraisalAdjustment" }
              }
            }
          },
          "buybox": {
            "type": "object",
            "properties": {
              "rehabLevelIndex": { "type": "integer", "minimum": 0, "maximum": 4, "default": 2, "description": "Rehab level 0-4 (0=as-is, 4=full gut)" },
              "majorItems": { "type": "array", "items": { "$ref": "#/components/schemas/MajorItem" } },
              "additionPlay": { "type": "number", "default": 0, "description": "Additional square footage value" },
              "closingCostsPercent": { "type": "number", "default": 10, "description": "Closing costs as % of ARV" },
              "carryingCostsPercent": { "type": "number", "default": 5, "description": "Carrying costs as % of ARV" },
              "wholesaleFee": { "type": "number", "default": 10000, "description": "Wholesale fee in dollars" },
              "desiredProfit": { "type": "number", "description": "Desired profit in dollars" }
            }
          },
          "enrichment": {
            "type": "object",
            "properties": {
              "permits": { "type": "boolean", "default": true, "description": "Include permit history" },
              "floodZone": { "type": "boolean", "default": true, "description": "Include flood zone data" },
              "weatherRisk": { "type": "boolean", "default": false, "description": "Include weather risk data" }
            }
          },
          "photoAnalysis": {
            "type": "object",
            "description": "Photo fetching and AI vision analysis options",
            "properties": {
              "enabled": { "type": "boolean", "default": true, "description": "Enable photo fetching and AI vision analysis" },
              "provider": { "type": "string", "enum": ["zillow", "mls", "redfin"], "description": "Photo provider (auto-detected if not specified)" },
              "maxComps": { "type": "integer", "default": 10, "description": "Max comps to fetch photos for" },
              "requireBetterOrEqual": { "type": "boolean", "default": true, "description": "Require comps to be better than or equal to subject" }
            }
          },
          "zillowContext": {
            "type": "object",
            "description": "Zillow URL context fetching via Gemini (provides photos, description, price history)",
            "properties": {
              "enabled": { "type": "boolean", "default": true, "description": "Enable Zillow context fetching" },
              "skipCache": { "type": "boolean", "default": false, "description": "Skip cache and fetch fresh data" },
              "maxComps": { "type": "integer", "default": 10, "description": "Max comps to fetch Zillow data for" }
            }
          },
          "skipCache": { "type": "boolean", "default": false, "description": "Skip cache and fetch fresh data from APIs" }
        }
      },
      "AppraisalFilter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["subdivision_match", "sale_age", "sqft_diff", "property_type", "year_built_diff", "distance"],
            "description": "Filter type"
          },
          "enabled": { "type": "boolean", "description": "Whether this filter is active" },
          "value": { "type": "number", "description": "Filter threshold value" }
        }
      },
      "AppraisalAdjustment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["old_comp_discount", "bedroom", "bathroom", "pool", "garage", "carport"],
            "description": "Adjustment type"
          },
          "enabled": { "type": "boolean", "description": "Whether this adjustment is active" },
          "amount": { "type": "number", "description": "Dollar amount adjustment" },
          "percent": { "type": "number", "description": "Percentage adjustment (for old_comp_discount)" }
        }
      },
      "MajorItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "enabled": { "type": "boolean" },
          "cost": { "type": "number" }
        }
      },
      "QueueJobResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "jobId": { "type": "string", "example": "job_1708123456789_abc1234" },
              "propertyKey": { "type": "string", "description": "Property key for status lookup" },
              "status": { "type": "string", "enum": ["queued"], "example": "queued" },
              "streamUrl": { "type": "string", "description": "SSE URL for real-time updates" },
              "pollUrl": { "type": "string", "description": "HTTP URL for polling status" },
              "estimatedDurationMs": { "type": "integer", "example": 15000, "description": "Estimated job duration in milliseconds" }
            }
          }
        }
      },
      "JobStatusResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "jobId": { "type": "string" },
              "status": { "type": "string", "enum": ["queued", "running", "completed", "failed"] },
              "progress": { "type": "number", "minimum": 0, "maximum": 100 },
              "currentStep": { "type": "string" },
              "result": { "type": "object", "description": "Analysis result (when completed)" },
              "error": { "type": "string", "description": "Error message (when failed)" }
            }
          }
        }
      },
      "AnalyzeDefaultsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "searchOptions": {
                "type": "object",
                "properties": {
                  "radiusMiles": { "type": "number", "example": 1 },
                  "maxComps": { "type": "integer", "example": 10 },
                  "monthsBack": { "type": "integer", "example": 12 }
                }
              },
              "appraisalRules": {
                "type": "object",
                "properties": {
                  "filters": { "type": "array", "items": { "$ref": "#/components/schemas/AppraisalFilter" } },
                  "adjustments": { "type": "array", "items": { "$ref": "#/components/schemas/AppraisalAdjustment" } }
                }
              },
              "buybox": {
                "type": "object",
                "properties": {
                  "rehabLevelIndex": { "type": "integer", "example": 2 },
                  "closingCostsPercent": { "type": "number", "example": 10 },
                  "carryingCostsPercent": { "type": "number", "example": 5 },
                  "wholesaleFee": { "type": "number", "example": 10000 }
                }
              },
              "enrichment": {
                "type": "object",
                "properties": {
                  "permits": { "type": "boolean", "example": true },
                  "floodZone": { "type": "boolean", "example": true },
                  "weatherRisk": { "type": "boolean", "example": false }
                }
              },
              "photoAnalysis": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean", "example": true },
                  "maxComps": { "type": "integer", "example": 10 },
                  "requireBetterOrEqual": { "type": "boolean", "example": true }
                }
              },
              "zillowContext": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean", "example": true },
                  "skipCache": { "type": "boolean", "example": false },
                  "maxComps": { "type": "integer", "example": 10 }
                }
              },
              "rehabLevels": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "index": { "type": "integer" },
                    "name": { "type": "string" }
                  }
                }
              },
              "majorItems": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "defaultCost": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}